Move EncryptorUtil to a new crypto package.
